Relocate Grid type definition to commonTypes directory
