+++ src/app/[locale]/dashboard/@board/Grid/types.ts
@@ -1,11 +1,3 @@
 export type Item = {
   id: string;
 };
-
-export type Grid = {
-  rows: number;
-  columns: number;
-  order: GridOrder;
-};
-
-export type GridOrder = (string | null)[][];
