+++ src/commonTypes/Grid.ts
@@ -1,0 +1,7 @@
+export type Grid = {
+  rows: number;
+  columns: number;
+  order: GridOrder;
+};
+
+export type GridOrder = (string | null)[][];
